リファクタしましたテスト動きません
