--- src/main/java/matcheam/StartTime.java
@@ -1,18 +1,0 @@
-package matcheam;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-/**
- * Created by ooguro on 2017/01/07.
- */
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class StartTime {
-
-	private LocalDateTime value;
-}
