+++ src/main/java/matcheam/match/GameTime.java
@@ -1,13 +1,17 @@
-package matcheam;
+package matcheam.match;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.time.LocalTime;
 
 /**
  * Created by ooguro on 2017/01/07.
  */
 @AllArgsConstructor
 @NoArgsConstructor
-public class Enemy {
-	private String value = "";
+@Getter
+public class GameTime {
+	private LocalTime value;
 }
