+++ src/main/java/matcheam/match/StartTime.java
@@ -1,10 +1,10 @@
-package matcheam;
+package matcheam.match;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalTime;
+import java.time.LocalDateTime;
 
 /**
  * Created by ooguro on 2017/01/07.
@@ -12,6 +12,7 @@ import java.time.LocalTime;
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class GameTime {
-	private LocalTime value;
+public class StartTime {
+
+	private LocalDateTime value;
 }
