+++ src/main/java/matcheam/person/Person.java
@@ -1,19 +1,21 @@
-package matcheam;
+package matcheam.person;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import matcheam.match.Identifier;
+import sun.security.provider.DSAPublicKeyImpl;
 
 import java.util.UUID;
 
 @Data
 @NoArgsConstructor
-public class EntryUser {
+public class Person {
 
 	private Identifier identifier;
 
 	private String name;
 
-	EntryUser(String name) {
+	public Person(String name) {
 		this.identifier = new Identifier(UUID.randomUUID().toString());
 		this.name = name;
 	}
