+++ src/main/java/matcheam/person/PersonService.java
@@ -1,5 +1,6 @@
-package matcheam;
+package matcheam.person;
 
+import matcheam.match.Identifier;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
@@ -8,17 +9,17 @@ import java.util.HashMap;
  * Created by ooguro on 2017/01/07.
  */
 @Service
-public class EntryUserService {
+public class PersonService {
 
 	// TODO 永続化する
-	public HashMap<String, EntryUser> entryUserHashMap = new HashMap<>();
+	public HashMap<String, Person> entryUserHashMap = new HashMap<>();
 
-	public void register(EntryUser entryUser) {
+	public void entry(Person entryUser) {
 		// TODO 永続化する
 		entryUserHashMap.put(entryUser.getIdentifier().toString(), entryUser);
 	}
 
-	public EntryUser findOne(Identifier identifier) {
+	public Person findOne(Identifier identifier) {
 		return entryUserHashMap.get(identifier.toString());
 	}
 }
