+++ src/test/java/matcheam/EntryUserServiceTest.java
@@ -1,5 +1,8 @@
 package matcheam;
 
+import matcheam.match.Identifier;
+import matcheam.person.Person;
+import matcheam.person.PersonService;
 import org.junit.Test;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
@@ -9,17 +12,17 @@ import static org.assertj.core.api.Java6Assertions.assertThat;
  */
 public class EntryUserServiceTest {
 
-	private EntryUserService entryUserService = new EntryUserService();
+	private PersonService entryUserService = new PersonService();
 
 	@Test
 	public void 登録した申込者を検索できること() throws Exception {
-		EntryUser entryUser = new EntryUser();
+		Person entryUser = new Person();
 		entryUser.setIdentifier(new Identifier("1"));
 		entryUser.setName("名前");
 
-		entryUserService.register(entryUser);
+		entryUserService.entry(entryUser);
 
-		EntryUser actual = entryUserService.entryUserHashMap.get("1");
+		Person actual = entryUserService.entryUserHashMap.get("1");
 		assertThat(actual.getName()).isEqualTo("名前");
 		assertThat(actual.getIdentifier().toString()).isEqualTo("1");
 	}
